[Longest Palindrome] Candidate Pairs Strategy
The focus of these changes is to implement the
solution to the challenge based off of Jesse's
idea of identifying candidate character pairs
that could form the basis of a palindrome.  This
is a straightforward implementation rather than a
fully optimized one, but it serves as a solid
foundation for further enhancements.